Refactored config dataset
